+++ database/migrations/2024_10_22_214143_add_kd_status_to_jadwal_and_tickets.php
@@ -11,15 +11,16 @@ return new class extends Migration
      */
     public function up()
     {
-    
-        // Menambahkan kolom kd_status di tabel tickets
-        Schema::table('tickets', function (Blueprint $table) {
-            $table->string('kd_status', 5)->nullable();
+        // Menambahkan kolom kd_status di tabel tickets jika belum ada
+        if (!Schema::hasColumn('tickets', 'kd_status')) {
+            Schema::table('tickets', function (Blueprint $table) {
+                $table->string('kd_status', 5)->nullable();
 
-            // Menambahkan foreign key constraint
-            $table->foreign('kd_status')->references('kd_status')->on('kategori_status')
-                  ->onUpdate('cascade')->onDelete('set null');
-        });
+                // Menambahkan foreign key constraint
+                $table->foreign('kd_status')->references('kd_status')->on('kategori_status')
+                      ->onUpdate('cascade')->onDelete('set null');
+            });
+        }
     }
 
     /**
@@ -27,16 +28,20 @@ return new class extends Migration
      */
     public function down()
     {
-        // Drop foreign key dan kolom jika di-rollback di tabel jadwal
-        Schema::table('jadwal', function (Blueprint $table) {
-            $table->dropForeign(['kd_status']);
-            $table->dropColumn('kd_status');
-        });
-
         // Drop foreign key dan kolom jika di-rollback di tabel tickets
         Schema::table('tickets', function (Blueprint $table) {
-            $table->dropForeign(['kd_status']);
-            $table->dropColumn('kd_status');
+            if (Schema::hasColumn('tickets', 'kd_status')) {
+                $table->dropForeign(['kd_status']);
+                $table->dropColumn('kd_status');
+            }
+        });
+
+        // Drop foreign key dan kolom jika di-rollback di tabel jadwal
+        Schema::table('jadwal', function (Blueprint $table) {
+            if (Schema::hasColumn('jadwal', 'kd_status')) {
+                $table->dropForeign(['kd_status']);
+                $table->dropColumn('kd_status');
+            }
         });
     }
 };
